requirements.txt
asgiref==3.6.0
attrs==22.2.0
bcrypt==4.0.1
black==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
defusedxml==0.7.1
distro==1.8.0
Django==4.1.7
django-extensions==3.2.1
django-phonenumber-field==7.0.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docker==6.0.1
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.21.5
flake8==6.0.0
idna==3.4
inflection==0.5.1
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
jsonschema==3.2.0
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.0
paramiko==3.1.0
pathspec==0.11.0
phonenumbers==8.13.7
Pillow==9.4.0
platformdirs==3.0.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
PyNaCl==1.5.0
pyrsistent==0.19.3
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2023.3
PyYAML==5.4.1
requests==2.28.2
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.1
sqlparse==0.4.3
texttable==1.6.7
tomli==2.0.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
websocket-client==0.59.0
